# docs/openapi-v1.yaml
openapi: 3.1.0
info:
  title: HireX Unified Candidate + Recruiter API
  version: 1.0.0-draft
  description: 'DRAFT API v1 contract for Candidate Android and Recruiter Web. HTTP Method + normalized Path is
    the unique operation key. JSON fields use camelCase; all IDs are opaque strings; all date-time values are ISO-8601
    UTC and end with Z. OpenAPI is the implementation source of truth. MVP decisions: SGD integer major units; one
    Resume per Candidate; Recruiter registration creates a company; fixed transition matrices; deferred features
    are marked with x-mvp-scope.'
servers:
- url: http://localhost:8080/api/v1
  description: MVP local Web and backend development
- url: http://10.0.2.2:8080/api/v1
  description: MVP Android Emulator access to the host backend
...
security:
- bearerAuth: []
paths:
  /auth/register:
    post:
      tags:
      - Auth
      summary: Register Candidate or Recruiter
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
      x-status: DRAFT
      x-clients:
      - Candidate
      - Recruiter
      x-permission: Public
      description: 'Permission: Public. Status: DRAFT. Clients: Candidate, Recruiter.'
      x-mvp-scope: MVP
  /auth/login:
    post:
      tags:
      - Auth
      summary: Log in with email and password
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
      x-status: DRAFT
      x-clients:
      - Candidate
      - Recruiter
      x-permission: Public
      description: 'Permission: Public. Status: DRAFT. Clients: Candidate, Recruiter.'
      x-mvp-scope: MVP
  /auth/refresh:
    post:
      tags:
      - Auth
      summary: Rotate access and refresh tokens
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Tokens refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
      x-status: DRAFT
      x-clients:
      - Candidate
      - Recruiter
      x-permission: Public
      description: 'Permission: Public. Status: DRAFT. Clients: Candidate, Recruiter.'
      x-mvp-scope: MVP
  /auth/logout:
    post:
      tags:
      - Auth
      summary: Revoke the current refresh token
      operationId: logout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '204':
          description: Logged out
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      x-status: DRAFT
      x-clients:
      - Candidate
      - Recruiter
      x-permission: Authenticated session
      description: 'Permission: Authenticated session. Status: DRAFT. Clients: Candidate, Recruiter.'
      x-mvp-scope: MVP
  /auth/password-reset/request:
    post:
      tags: [Auth]
      summary: Request a one-time password reset code
      operationId: requestPasswordReset
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PasswordResetRequest' }
      responses:
        '202':
          description: The same accepted response is returned whether or not the email exists.
        '422': { $ref: '#/components/responses/ValidationError' }
        '503':
          description: Email delivery is not configured or unavailable.
      x-permission: Public
      x-clients: [Candidate, Recruiter]
  /auth/password-reset/confirm:
    post:
      tags: [Auth]
      summary: Confirm a one-time code and replace the password
      operationId: confirmPasswordReset
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PasswordResetConfirmRequest' }
      responses:
        '204': { description: Password changed and all existing sessions revoked. }
        '422': { $ref: '#/components/responses/ValidationError' }
      x-permission: Public
      x-clients: [Candidate, Recruiter]
...
  /community/posts:
    get:
      tags: [Community]
      summary: List the authenticated community feed
      operationId: listCommunityPosts
      security:
      - bearerAuth: []
      parameters:
      - name: q
        in: query
        schema: {type: string}
      - name: category
        in: query
        schema: {$ref: '#/components/schemas/CommunityCategory'}
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/CommunityPageSize'
      responses:
        '200':
          description: Community posts ordered by createdAt descending, then id descending
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CommunityFeedResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
      x-status: DRAFT
      x-clients: [Candidate, Recruiter]
      x-permission: Authenticated Candidate or Recruiter only
      description: 'Permission: authenticated CANDIDATE or RECRUITER only. Stable sort: createdAt DESC, then id DESC. Counts and likedByCurrentUser are server-derived for the authenticated viewer.'
      x-mvp-scope: P2_DEMO
    post:
      tags: [Community]
      summary: Create a categorized community post with up to four images
      operationId: createCommunityPost
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateCommunityPostRequest' }
          multipart/form-data:
            schema:
              type: object
              required: [body, category]
              properties:
                body: {type: string, minLength: 1, maxLength: 2000}
                category: {$ref: '#/components/schemas/CommunityCategory'}
                images:
                  type: array
                  maxItems: 4
                  items: {type: string, format: binary}
      responses:
        '201':
          description: Community post created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CommunityPostResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
      x-status: DRAFT
      x-clients: [Candidate, Recruiter]
      x-permission: Authenticated Candidate or Recruiter only
      description: 'Permission: authenticated CANDIDATE or RECRUITER only. Text-only body is trimmed by the server and must contain 1 to 2,000 characters after trimming.'
      x-mvp-scope: P2_DEMO
  /community/posts/{postId}/images/{imageId}:
    get:
      tags: [Community]
      summary: Download a Community post image
      operationId: getCommunityPostImage
      security: []
      parameters:
      - $ref: '#/components/parameters/CommunityPostId'
      - name: imageId
        in: path
        required: true
        schema: {type: string}
      responses:
        '200':
          description: PNG, JPEG, or WebP image bytes
          content:
            image/png: {schema: {type: string, format: binary}}
            image/jpeg: {schema: {type: string, format: binary}}
            image/webp: {schema: {type: string, format: binary}}
        '404': {$ref: '#/components/responses/NotFound'}
...
  /community/posts/{postId}:
    get:
      tags: [Community]
      summary: Get a community post
      operationId: getCommunityPost
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/CommunityPostId'
      responses:
        '200':
          description: Community post detail without embedded comments
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CommunityPostResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
      x-status: DRAFT
      x-clients: [Candidate, Recruiter]
      x-permission: Authenticated Candidate or Recruiter only
      description: 'Permission: authenticated CANDIDATE or RECRUITER only. Returns 404 when the Post does not exist. Comments are fetched through the separate comments operation.'
      x-mvp-scope: P2_DEMO
  /community/posts/{postId}/like:
    put:
      tags: [Community]
      summary: Like a community post
      operationId: likeCommunityPost
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/CommunityPostId'
      responses:
        '200':
          description: Latest server-derived Like state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CommunityInteractionResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
      x-status: DRAFT
      x-clients: [Candidate, Recruiter]
      x-permission: Authenticated Candidate or Recruiter only
      description: 'Permission: authenticated CANDIDATE or RECRUITER only. Idempotent: liking an already-liked Post succeeds without creating another Like. Returns the latest likeCount and likedByCurrentUser=true.'
      x-idempotent: true
      x-mvp-scope: P2_DEMO
    delete:
      tags: [Community]
      summary: Remove the current user's Like from a community post
      operationId: unlikeCommunityPost
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/CommunityPostId'
      responses:
        '200':
          description: Latest server-derived Like state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CommunityInteractionResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
      x-status: DRAFT
      x-clients: [Candidate, Recruiter]
      x-permission: Authenticated Candidate or Recruiter only
      description: 'Permission: authenticated CANDIDATE or RECRUITER only. Idempotent: removing a Like that does not exist succeeds. Returns the latest likeCount and likedByCurrentUser=false.'
      x-idempotent: true
      x-mvp-scope: P2_DEMO
  /community/posts/{postId}/comments:
    get:
      tags: [Community]
      summary: List comments for a community post
      operationId: listCommunityComments
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/CommunityPostId'
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/CommunityPageSize'
      responses:
        '200':
          description: Comments ordered by createdAt ascending, then id ascending
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CommunityCommentListResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
      x-status: DRAFT
      x-clients: [Candidate, Recruiter]
      x-permission: Authenticated Candidate or Recruiter only
      description: 'Permission: authenticated CANDIDATE or RECRUITER only. Returns 404 when the Post does not exist. Stable sort: createdAt ASC, then id ASC.'
      x-mvp-scope: P2_DEMO
    post:
      tags: [Community]
      summary: Create a text-only comment on a community post
      operationId: createCommunityComment
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/CommunityPostId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateCommunityCommentRequest' }
      responses:
        '201':
          description: Community comment created with the Post's latest comment count
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreateCommunityCommentResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
      x-status: DRAFT
      x-clients: [Candidate, Recruiter]
      x-permission: Authenticated Candidate or Recruiter only
      description: 'Permission: authenticated CANDIDATE or RECRUITER only. Text-only body is trimmed by the server and must contain 1 to 500 characters after trimming. Returns 404 when the Post does not exist.'
      x-mvp-scope: P2_DEMO
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
...
    CommunityPageSize:
      name: pageSize
      in: query
      description: Community-only page size. This does not change the shared PageSize maximum of 100.
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 20
    CommunityPostId:
      name: postId
      in: path
      required: true
      description: Opaque Community Post identifier.
      schema:
        type: string
...
  responses:
    Unauthorized:
      description: Missing, invalid, or expired credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated principal has the wrong role or insufficient capability
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    CommunityPost:
      type: object
      additionalProperties: false
      required: [id, author, body, category, images, likeCount, commentCount, likedByCurrentUser, createdAt, updatedAt]
      properties:
        id: { type: string }
        author: { $ref: '#/components/schemas/CommunityAuthor' }
        body: { type: string }
        category: {$ref: '#/components/schemas/CommunityCategory'}
        images:
          type: array
          maxItems: 4
          items: {$ref: '#/components/schemas/CommunityImage'}
        likeCount: { type: integer, minimum: 0 }
        commentCount: { type: integer, minimum: 0 }
        likedByCurrentUser: { type: boolean }
        createdAt:
          type: string
          format: date-time
          pattern: ^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(?:\.\d+)?Z$
          description: ISO-8601 UTC date-time; the value MUST end with Z.
        updatedAt:
          type: string
          format: date-time
          pattern: ^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(?:\.\d+)?Z$
          description: ISO-8601 UTC date-time; the value MUST end with Z.
    CommunityComment:
      type: object
      additionalProperties: false
      required: [id, postId, author, body, createdAt, updatedAt]
      properties:
        id: { type: string }
        postId: { type: string }
        author: { $ref: '#/components/schemas/CommunityAuthor' }
        body: { type: string }
        createdAt:
          type: string
          format: date-time
          pattern: ^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(?:\.\d+)?Z$
          description: ISO-8601 UTC date-time; the value MUST end with Z.
        updatedAt:
          type: string
          format: date-time
          pattern: ^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(?:\.\d+)?Z$
          description: ISO-8601 UTC date-time; the value MUST end with Z.
    CreateCommunityPostRequest:
      type: object
      additionalProperties: false
      required: [body, category]
      properties:
        body:
          type: string
          minLength: 1
          maxLength: 2000
          pattern: ^.*\S.*$
          description: Text only. The server trims leading and trailing whitespace before validation; the normalized value must contain 1 to 2,000 characters and at least one non-whitespace character.
          x-normalization: trim
          x-trimmed-minLength: 1
          x-trimmed-maxLength: 2000
        category: {$ref: '#/components/schemas/CommunityCategory'}
    CommunityCategory:
      type: string
      enum: [JOB_SEEKING, RECRUITING, TECH_DISCUSSION, HELP, GENERAL]
    CommunityImage:
      type: object
      required: [imageId, url, contentType, sizeBytes]
      properties:
        imageId: {type: string}
        url: {type: string}
        contentType: {type: string, enum: [image/png, image/jpeg, image/webp]}
        sizeBytes: {type: integer, minimum: 1, maximum: 5242880}
...
    CreateCommunityCommentRequest:
      type: object
      additionalProperties: false
      required: [body]
      properties:
        body:
          type: string
          minLength: 1
          maxLength: 500
          pattern: ^.*\S.*$
          description: Text only. The server trims leading and trailing whitespace before validation; the normalized value must contain 1 to 500 characters and at least one non-whitespace character.
          x-normalization: trim
          x-trimmed-minLength: 1
          x-trimmed-maxLength: 500
    CommunityInteraction:
      type: object
      additionalProperties: false
      required: [postId, likeCount, likedByCurrentUser]
      properties:
        postId: { type: string }
        likeCount: { type: integer, minimum: 0 }
        likedByCurrentUser: { type: boolean }
    CommunityFeedResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/CommunityPost' }
        meta: { $ref: '#/components/schemas/PageMeta' }
    CommunityCommentListResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/CommunityComment' }
        meta: { $ref: '#/components/schemas/PageMeta' }
    CommunityPostResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/CommunityPost' }
    CommunityInteractionResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/CommunityInteraction' }
    CreateCommunityCommentResult:
      type: object
      additionalProperties: false
      required: [comment, commentCount]
      properties:
        comment: { $ref: '#/components/schemas/CommunityComment' }
        commentCount:
          type: integer
          minimum: 0
          description: Latest server-derived comment count for the Post after creation.
    CreateCommunityCommentResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/CreateCommunityCommentResult' }
    RegisterRequest:
      oneOf:
      - $ref: '#/components/schemas/CandidateRegisterRequest'
      - $ref: '#/components/schemas/RecruiterRegisterRequest'
      discriminator:
        propertyName: role
        mapping:
          CANDIDATE: '#/components/schemas/CandidateRegisterRequest'
          RECRUITER: '#/components/schemas/RecruiterRegisterRequest'
    LoginRequest:
      type: object
      additionalProperties: false
      required:
      - email
      - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    PasswordResetRequest:
      type: object
      additionalProperties: false
      required: [email]
      properties:
        email: { type: string, format: email, maxLength: 255 }
    PasswordResetConfirmRequest:
      type: object
      additionalProperties: false
      required: [email, code, newPassword]
      properties:
        email: { type: string, format: email, maxLength: 255 }
        code: { type: string, pattern: '^\\d{6}$' }
        newPassword: { type: string, format: password, minLength: 8, maxLength: 128 }
...
    RefreshTokenRequest:
      type: object
      additionalProperties: false
      required:
      - refreshToken
      properties:
        refreshToken:
          type: string
          description: Single-use rotating refresh token. A successful refresh revokes it and returns a new refresh
            token.
    AuthUser:
      allOf:
      - $ref: '#/components/schemas/User'
      - type: object
        properties:
          company:
            oneOf:
            - $ref: '#/components/schemas/Company'
            - type: 'null'
        description: company is present for Recruiter accounts and null for Candidate and Admin accounts.
    TokenData:
      type: object
      required:
      - accessToken
      - refreshToken
      - expiresIn
      - refreshExpiresIn
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          const: 7200
          description: 'Access token lifetime in seconds: 2 hours for MVP.'
        refreshExpiresIn:
          type: integer
          const: 2592000
          description: 'Refresh token lifetime in seconds: 30 days for MVP.'
    AuthResponse:
      type: object
      required:
      - data
      properties:
        data:
          allOf:
          - $ref: '#/components/schemas/TokenData'
          - type: object
            required:
            - user
            properties:
              user:
                $ref: '#/components/schemas/AuthUser'
              onboardingRequired:
                type: [boolean, 'null']
                description: Present for Candidate accounts; omitted or null for Recruiter accounts.
    TokenResponse:
      type: object
      required:
      - data
      properties:
        data:
          $ref: '#/components/schemas/TokenData'
...
    PageMeta:
      type: object
      required:
      - page
      - pageSize
      - total
      - hasNext
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        hasNext:
          type: boolean
...
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
...
    CandidateRegisterRequest:
      type: object
      additionalProperties: false
      required:
      - role
      - fullName
      - email
      - password
      - acceptedTermsVersion
      properties:
        role:
          type: string
          const: CANDIDATE
        fullName:
          type: string
          minLength: 1
          maxLength: 100
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 128
        acceptedTermsVersion:
          type: string
    RecruiterRegisterRequest:
      type: object
      additionalProperties: false
      required:
      - role
      - companyName
      - fullName
      - email
      - password
      - acceptedTermsVersion
      properties:
        role:
          type: string
          const: RECRUITER
        companyName:
          type: string
          minLength: 1
          maxLength: 200
        fullName:
          type: string
          minLength: 1
          maxLength: 100
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 128
        acceptedTermsVersion:
          type: string
      description: 'MVP: registration creates a new Company and makes this Recruiter the company''s only administrator.
        Joining an existing company and invitations are deferred.'
...
    User:
      type: object
      required:
      - userId
      - role
      - fullName
      - email
      - createdAt
      - updatedAt
      properties:
        userId:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        fullName:
          type: string
        email:
          type: string
          format: email
        avatarUrl:
          oneOf:
          - type: string
            format: uri
          - type: 'null'
        createdAt:
          type: string
          format: date-time
          pattern: ^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(?:\.\d+)?Z$
          description: ISO-8601 UTC date-time; the value MUST end with Z.
        updatedAt:
          type: string
          format: date-time
          pattern: ^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(?:\.\d+)?Z$
          description: ISO-8601 UTC date-time; the value MUST end with Z.
      description: Canonical account identity shared by Candidate and Recruiter.
...